requirements.txt
PyYAML==5.4.1
packaging==23.1
imutils
numpy
grpcio==1.56.0
opencv-python-headless==4.8.0.74
protobuf==4.23.4
shapely==2.0.1
pyclipper==1.3.0.post4
redis==4.6.0
unidecode==1.3.6
python-Levenshtein==0.21.1
tritonclient==2.35.0
bpemb